requirements.txt
transformers==5.5.4
openai==2.32.0
detoxify==0.5.2
requests==2.33.1
streamlit==1.56.0
python-dotenv==1.2.2